web_mapping/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-cors-headers==3.13.0
django-countries==7.5.1
djangorestframework==3.14.0
Pillow==9.4.0
psycopg2==2.9.5
pytz==2022.7
sqlparse==0.4.3
typing_extensions==4.6.3
